fix(deps): update rust crate tree-sitter to 0.27 (#503)
* fix(deps): update rust crate tree-sitter to 0.27

* fix: adapt to tree-sitter 0.27 node API changes

tree-sitter 0.27 makes two breaking changes that this bump did not account
for, so the branch failed to compile.

QueryMatch::captures is now a private field with a captures() accessor
returning the same slice, so the two iteration sites in flags.rs call it.

Node::child_count() now returns u32 rather than usize, and Node::child()
takes u32, so the 40 `as u32` casts that converted the old usize return are
dead. Removing them is mechanical and changes no behavior; clippy rejects
them under -D warnings as casts to the same type.


---------

+++ src/analyzers/deadcode.rs
@@ -1092,7 +1092,7 @@ fn collect_function_value_refs(
     let args_node = call_node.child_by_field_name("arguments").or_else(|| {
         // Some grammars (Go) use "argument_list" as the node kind
         // rather than a named field
-        (0..call_node.child_count() as u32)
+        (0..call_node.child_count())
             .filter_map(|i| call_node.child(i))
             .find(|c| c.kind() == "argument_list")
     });
@@ -1102,7 +1102,7 @@ fn collect_function_value_refs(
         None => return,
     };
 
-    for i in 0..args_node.child_count() as u32 {
+    for i in 0..args_node.child_count() {
         let arg = match args_node.child(i) {
             Some(a) => a,
             None => continue,
